Collecting spectroscopy routines into AggregateSpectroscopy class
