+++ setup.py
@@ -27,7 +27,7 @@ setup(
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='0.0.32',
+    version='0.0.33',
 
     description='Quantarhei: Open Quantum System Theory for Molecular Systems',
     long_description=long_description,
